+++ src/trunk/base/util-safe.c
@@ -64,7 +64,7 @@ _smalloc(size_t sz, const char * file, int line)
     v = malloc(sz);
     if (v == NULL) {
         fprintf(stderr, "[%2s]  **** PANIC: malloc %u bytes (%s:%d): %s\n",
-            getprocname(map.whoami), sz, file, line, strerror(errno));
+		getprocname(map.whoami), (unsigned int)sz, file, line, strerror(errno));
         abort();
     }
 
@@ -90,7 +90,8 @@ _scalloc(size_t n, size_t sz, const char * file, int line)
     v = calloc(n, sz);
     if (v == NULL) {
 	fprintf(stderr, "[%2s]  **** PANIC: calloc n:%u sz:%u (%s:%d): %s\n", 
-	    getprocname(map.whoami), n, sz, file, line, strerror(errno));
+		getprocname(map.whoami), (unsigned int)n, (unsigned int)sz, 
+		file, line, strerror(errno));
 	abort();
     }
